Plot a Petri net as a bipartite graph via DiagrammeR

Species are shown as blue circles, transitions as orange boxes. Accepts LabelledPetriNet, TypedPetriNet, or Open Petri nets.

Usage

plot_petri(pn)

Arguments

pn

A Petri net ACSet, TypedPetriNet, or Open Petri net

Value

A DiagrammeR htmlwidget

Examples

sir <- labelled_petri_net(
  c("S", "I", "R"),
  inf = c("S", "I") %=>% c("I", "I"),
  rec = "I" %=>% "R"
)
if (requireNamespace("DiagrammeR", quietly = TRUE)) {
  plot_petri(sir)
}